change method sig to accept Long

+++ src/main/java/com/artronics/security/JWTLoginFilter.java
@@ -67,7 +67,7 @@ public class JWTLoginFilter extends AbstractAuthenticationProcessingFilter {
 
         addUserAndAccountToRes(res, auth);
 
-        TokenAuthenticationService.addAuthentication(res, auth.getPrincipal());
+        TokenAuthenticationService.addAuthentication(res, (Long) auth.getPrincipal());
     }
 
     private void addUserAndAccountToRes(HttpServletResponse res, Authentication auth) throws IOException {
